+++ src/routes.js
@@ -8,14 +8,11 @@ import Results from "layouts/results";
 
 // @mui icons
 import DashboardRoundedIcon from "@mui/icons-material/DashboardRounded";
-// import GroupAddIcon from "@mui/icons-material/GroupAdd";
 import GroupsRoundedIcon from "@mui/icons-material/GroupsRounded";
 import BorderColorRoundedIcon from "@mui/icons-material/BorderColorRounded";
 import SchoolIcon from "@mui/icons-material/School";
 import LibraryBooksRoundedIcon from "@mui/icons-material/LibraryBooksRounded";
 import AssignmentTurnedInRoundedIcon from "@mui/icons-material/AssignmentTurnedInRounded";
-
-// import EditNoteRoundedIcon from '@mui/icons-material/EditNoteRounded';
 
 const routes = [
   {
